fix: change column type in database

+++ database/migrations/1745340376700_create_transcription_parts_table.ts
@@ -8,7 +8,7 @@ export default class extends BaseSchema {
       table.increments('id')
 
       table.string('discord_user_id').nullable()
-      table.integer('start_time').notNullable()
+      table.float('start_time').notNullable()
       table.float('duration').notNullable()
       table.integer('meeting_id').references('id').inTable('meetings').notNullable()
       table.text('text')
